pull apart tests a bit; docs for new functions; don't need to override new property
